refactor: Simplify builds
Simplify builds and entrypoints
