+++ jsr.json
@@ -1,6 +1,6 @@
 {
   "name": "@kasperrt/wiretyped",
-  "version": "0.2.3",
+  "version": "0.2.4-rc-1",
   "exports": {
     ".": "./src/index.ts"
   },
