+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wiretyped",
-  "version": "0.2.3",
+  "version": "0.2.4-rc-1",
   "description": "Typed HTTP client with error-first ergonomics, caching, retries, SSE, and runtime validation powered by @standard-schema/spec.",
   "type": "module",
   "keywords": [
@@ -26,39 +26,14 @@
   "packageManager": "pnpm@10.20.0",
   "main": "./dist/index.cjs",
   "module": "./dist/index.mjs",
-  "types": "./dist/types/index.d.mts",
+  "types": "./dist/types/index.d.ts",
   "files": ["dist"],
   "sideEffects": false,
   "exports": {
     ".": {
-      "import": {
-        "types": "./dist/types/index.d.mts",
-        "default": "./dist/index.mjs"
-      },
-      "require": {
-        "types": "./dist/types/index.d.cts",
-        "default": "./dist/index.cjs"
-      }
-    },
-    "./core": {
-      "import": {
-        "types": "./dist/types/core/index.d.mts",
-        "default": "./dist/core.mjs"
-      },
-      "require": {
-        "types": "./dist/types/core/index.d.cts",
-        "default": "./dist/core.cjs"
-      }
-    },
-    "./error": {
-      "import": {
-        "types": "./dist/types/error/index.d.mts",
-        "default": "./dist/error.mjs"
-      },
-      "require": {
-        "types": "./dist/types/error/index.d.cts",
-        "default": "./dist/error.cjs"
-      }
+      "types": "./dist/types/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
     },
     "./package.json": "./package.json"
   },
@@ -67,7 +42,7 @@
   },
   "scripts": {
     "build": "vite build && pnpm run build:types",
-    "build:types": "tsc --project tsconfig.build.json --emitDeclarationOnly && node scripts/generate-cts-types.mjs",
+    "build:types": "tsc --project tsconfig.build.json --emitDeclarationOnly",
     "check": "pnpm check:lint && pnpm check:format && pnpm check:ts",
     "check:format": "pnpm exec biome format --error-on-warnings",
     "check:lint": "pnpm exec biome lint --error-on-warnings",
