+++ tsconfig.build.json
@@ -4,7 +4,7 @@
     "noEmit": false,
     "emitDeclarationOnly": true,
     "declaration": true,
-    "declarationMap": true,
+    "declarationMap": false,
     "outDir": "./dist/types",
     "rootDir": "./src"
   },
