+++ vite.config.ts
@@ -8,11 +8,7 @@ const enableBundle = process.env.CODECOV_BUNDLE_ANALYSIS === 'true';
 export default defineConfig({
   build: {
     lib: {
-      entry: {
-        index: resolve(__dirname, 'src/index.ts'),
-        core: resolve(__dirname, 'src/core/index.ts'),
-        error: resolve(__dirname, 'src/error/index.ts'),
-      },
+      entry: resolve(__dirname, 'src/index.ts'),
     },
     outDir: 'dist',
     rolldownOptions: {
